added validation, simple security layer and performed convenient testing from main class

+++ src/main/java/com/krasnopolskyi/facade/MainFacade.java
@@ -4,22 +4,22 @@ import com.krasnopolskyi.dto.request.*;
 import com.krasnopolskyi.dto.response.TraineeResponseDto;
 import com.krasnopolskyi.dto.response.TrainerResponseDto;
 import com.krasnopolskyi.dto.response.TrainingResponseDto;
+import com.krasnopolskyi.entity.User;
+import com.krasnopolskyi.exception.AccessException;
 import com.krasnopolskyi.exception.EntityException;
 import com.krasnopolskyi.exception.GymException;
 import com.krasnopolskyi.exception.ValidateException;
+import com.krasnopolskyi.security.AuthenticationManager;
 import com.krasnopolskyi.service.TraineeService;
 import com.krasnopolskyi.service.TrainerService;
 import com.krasnopolskyi.service.TrainingService;
+import com.krasnopolskyi.service.UserService;
 import com.krasnopolskyi.validation.CommonValidator;
-import com.krasnopolskyi.validation.Create;
-import jakarta.validation.ConstraintViolation;
+import com.krasnopolskyi.validation.group.Create;
 import jakarta.validation.ConstraintViolationException;
-import jakarta.validation.Valid;
-import jakarta.validation.Validator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import org.springframework.validation.annotation.Validated;
 
 import java.util.*;
 
@@ -31,6 +31,32 @@ public class MainFacade {
     private final TrainerService trainerService;
     private final TrainingService trainingService;
     private final CommonValidator validator;
+    private final AuthenticationManager manager;
+    private final UserService userService;
+
+    public boolean changePassword(String password, String executor) {
+        try {
+            manager.checkPermissions(executor);
+            userService.changePassword(executor, password);
+            log.info("Password has been changed");
+            return true;
+        } catch (GymException e) {
+            log.warn(e.getMessage());
+        }
+        return false;
+    }
+
+    public boolean changeActivityStatus(String target, String executor) {
+        try {
+            manager.checkPermissions(executor);
+            User user = userService.changeActivityStatus(target);
+            log.info("User " + user.getUsername()+ " is " + (user.getIsActive() ? "active" : "inactive"));
+            return true;
+        } catch (GymException e) {
+            log.warn(e.getMessage());
+        }
+        return false;
+    }
 
     public TraineeResponseDto createTrainee(TraineeDto traineeDto) {
         try {
@@ -47,54 +73,52 @@ public class MainFacade {
         }
     }
 
-    public TraineeResponseDto findTraineeById(Long id) {
+    public TraineeResponseDto findTraineeByUsername(String username, String executor) {
         try {
-            TraineeResponseDto trainee = traineeService.findById(id);
-            log.info("trainee with " + id + " has been found");
-            return trainee;
-        } catch (EntityException e) {
-            log.info("Failed find trainee " + e.getMessage());
-            return null;
-        }
-    }
-
-    public TraineeResponseDto findTraineeByUsername(String username) {
-        try {
+            manager.checkPermissions(executor);
             TraineeResponseDto trainee = traineeService.findByUsername(username);
             log.info("trainee with " + username + " has been found");
             return trainee;
         } catch (EntityException e) {
-            log.info("Failed find trainee " + e.getMessage());
-            return null;
+            log.warn("Failed find trainee " + e.getMessage());
+        } catch (AccessException e) {
+            log.warn(e.getMessage());
         }
+        return null;
     }
 
-    public TraineeResponseDto updateTrainee(TraineeDto traineeDto) {
+    public TraineeResponseDto updateTrainee(TraineeDto traineeDto, String executor) {
         try {
+            manager.checkPermissions(executor);
             validator.validate(traineeDto, Create.class);
             TraineeResponseDto refreshedTrainee = traineeService.update(traineeDto);
             log.info("Trainee successfully refreshed");
             return refreshedTrainee;
         } catch (ConstraintViolationException e) {
             log.warn("Validation failed: " + e.getMessage());
-            return null;
+        } catch (AccessException e) {
+            log.warn(e.getMessage());
         } catch (GymException e) {
-            log.info("Failed update trainee " + traineeDto.getId());
-            throw null;
+            log.warn("Failed update trainee " + traineeDto.getId());
         }
+        return null;
     }
 
-    public boolean deleteTrainee(String username) {
+    public boolean deleteTrainee(String username, String executor) {
         try {
+            manager.checkPermissions(executor);
             boolean isDeleted = traineeService.delete(username);
             log.info("trainee with " + username + " has been deleted");
             return isDeleted;
         } catch (EntityException e) {
-            log.info("Failed attempt to delete trainee " + e.getMessage());
+            log.warn("Failed attempt to delete trainee " + e.getMessage());
+            return false;
+        } catch (AccessException e) {
+            log.warn(e.getMessage());
             return false;
         }
     }
-    //////////////////////////////////////////////////////////
+    ///////////////////////trainer///////////////////////////////////
 
     public TrainerResponseDto createTrainer(TrainerDto trainerDto) {
         try {
@@ -104,43 +128,47 @@ public class MainFacade {
             return trainer;
         } catch (ConstraintViolationException e) {
             log.warn("Validation failed: " + e.getMessage());
-            return null;
         } catch (GymException e) {
-            // here should be ExceptionHandler
             log.warn("Registration failed " + e.getMessage());
-            return null;
         }
+        return null;
     }
 
-    public TrainerResponseDto findTrainerById(Long id) {
+    public TrainerResponseDto updateTrainer(TrainerDto trainerDto, String executor) {
         try {
-            TrainerResponseDto maybeTrainer = trainerService.findById(id);
-            log.info("trainee with " + id + " has been found");
-            return maybeTrainer;
-        } catch (EntityException e) {
-            log.info("Failed find trainee " + e.getMessage());
-            return null;
-        }
-    }
-
-    public TrainerResponseDto updateTrainer(TrainerDto trainerDto) {
-        try {
+            manager.checkPermissions(executor);
             validator.validate(trainerDto, Create.class);
             TrainerResponseDto updatedTrainer = trainerService.update(trainerDto);
             log.info("Trainer successfully refreshed");
             return updatedTrainer;
         } catch (ConstraintViolationException e) {
             log.warn("Validation failed: " + e.getMessage());
-            return null;
+        } catch (AccessException e) {
+            log.warn(e.getMessage());
         } catch (GymException e) {
-            log.info("Failed update trainee " + trainerDto.getId());
-            return null;
+            log.warn("Failed update trainee " + trainerDto.getId() + " " + e.getMessage());
         }
+        return null;
+    }
+
+    public TrainerResponseDto findTrainerByUsername(String username, String executor) {
+        try {
+            manager.checkPermissions(executor);
+            TrainerResponseDto trainer = trainerService.findByUsername(username);
+            log.info("trainer with " + username + " has been found");
+            return trainer;
+        } catch (EntityException e) {
+            log.warn("Failed find trainer " + e.getMessage());
+        } catch (AccessException e) {
+            log.warn(e.getMessage());
+        }
+        return null;
     }
     //////////////////////////////////////////////////////////////
 
-    public TrainingResponseDto addTraining(TrainingDto trainingDto) {
+    public TrainingResponseDto addTraining(TrainingDto trainingDto, String executor) {
         try {
+            manager.checkPermissions(executor);
             validator.validate(trainingDto, Create.class);
             TrainingResponseDto training = trainingService.save(trainingDto);
             log.info("Trainining saved successfully");
@@ -149,51 +177,49 @@ public class MainFacade {
             log.warn("Validation failed: " + e.getMessage());
             return null;
         } catch (ValidateException e) {
-            // here should be ExceptionHandler
             log.warn("adding training session failed " + e.getMessage());
-            return null;
+        } catch (AccessException e) {
+            log.warn(e.getMessage());
         }
+        return null;
     }
 
-    public TrainingResponseDto findTrainingById(Long id) {
+    public List<TrainingResponseDto> getAllTrainingsByUsernameAndFilter(TrainingFilterDto filterDto, String executor) {
         try {
-            TrainingResponseDto trainingResponseDto = trainingService.findById(id);
-            log.info("training with " + id + " has been found");
-            return trainingResponseDto;
-        } catch (EntityException e) {
-            log.info("Failed find training " + e.getMessage());
-            return null;
-        }
-    }
-
-    public List<TrainingResponseDto> getAllTrainingsByUsernameAndFilter(TrainingFilterDto filterDto) {
-        try {
+            manager.checkPermissions(executor);
             validator.validate(filterDto, Create.class);
             return trainingService.getFilteredTrainings(filterDto);
         } catch (ConstraintViolationException e) {
             log.warn("Validation failed: " + e.getMessage());
-            return null;
         } catch (ValidateException e) {
-            log.info("Something went wrong " + e.getMessage());
-            return new ArrayList<>();
+            log.warn("Something went wrong " + e.getMessage());
+        } catch (AccessException e) {
+            log.warn(e.getMessage());
         }
+        return new ArrayList<>();
     }
 
-    public List<TrainerResponseDto> getAllNotAssignedTrainersByTraineeUsername(String username) {
+    public List<TrainerResponseDto> getAllNotAssignedTrainersByTraineeUsername(String username, String executor) {
         try {
+            manager.checkPermissions(executor);
             return traineeService.findAllNotAssignedTrainersByTrainee(username);
         } catch (EntityException e) {
-            log.info("Something went wrong " + e.getMessage());
-            return new ArrayList<>();
+            log.warn("Something went wrong " + e.getMessage());
+        } catch (AccessException e) {
+            log.warn(e.getMessage());
         }
+        return new ArrayList<>();
     }
 
-    public List<TrainerResponseDto> updateTrainers(TraineeDto trainee, List<TrainerDto> newTrainers) {
+    public List<TrainerResponseDto> updateTrainers(TraineeDto trainee, List<TrainerDto> newTrainers, String executor) {
         try {
+            manager.checkPermissions(executor);
             return traineeService.updateTrainers(trainee, newTrainers);
         } catch (EntityException e) {
-            log.info("Could not update trainers list " + e.getMessage());
-            return new ArrayList<>();
+            log.warn("Could not update trainers list " + e.getMessage());
+        } catch (AccessException e) {
+            log.warn(e.getMessage());
         }
+        return new ArrayList<>();
     }
 }
